fix: shifted websocket logic into seperate component in find-match
