+++ peerprep/package.json
@@ -16,6 +16,7 @@
     "@mantine/notifications": "^7.13.2",
     "@monaco-editor/react": "^4.6.0",
     "@tabler/icons-react": "^3.19.0",
+    "axios": "^1.7.7",
     "bootstrap-icons": "^1.11.3",
     "mantine-form-zod-resolver": "^1.1.0",
     "monaco-editor": "^0.52.0",
