+++ peerprep/src/components/signup.tsx
@@ -66,5 +66,4 @@ export default function SignupForm() {
       </Stack>
     </form>
   )
-
 }
